Switch to long

+++ steam-config-reader/src/test/java/com/selesse/steam/TestGames.java
@@ -25,7 +25,7 @@ public enum TestGames {
         this.gameId = gameId;
     }
 
-    public int getGameId() {
+    public long getGameId() {
         return gameId;
     }
